+++ pkg/cniplugin/networking.go
@@ -1,13 +1,16 @@
 package cniplugin
 
 import (
+	"fmt"
 	"net"
 
+	currentcni "github.com/containernetworking/cni/pkg/types/040"
 	"github.com/jboelensns/openstack-cni/pkg/util"
 	"github.com/vishvananda/netlink"
 	"github.com/vishvananda/netns"
 )
 
+// NetworkInterface represents a local network interface (e.g ens3)
 type NetworkInterface struct {
 	Name     string
 	DestName string
@@ -16,6 +19,7 @@ type NetworkInterface struct {
 
 //go:generate moq -pkg mocks -out ../fixtures/mocks/cniplugin_mocks.go . Networking
 
+// Networking provides the ability to manipulate a network interface
 type Networking interface {
 	Configure(namespace string, iface *NetworkInterface) error
 }
@@ -24,10 +28,12 @@ type networking struct {
 	nl util.NetlinkWrapper
 }
 
+// NewNetworking returns a new Networking
 func NewNetworking(nl util.NetlinkWrapper) *networking {
 	return &networking{nl: nl}
 }
 
+// Configure moves an existing network interface into a new network namespace with the provided IP address and name
 func (me *networking) Configure(namespace string, iface *NetworkInterface) error {
 	// Find the link by interface name
 	link, err := me.nl.LinkByName(iface.Name)
@@ -81,4 +87,42 @@ func (me *networking) Configure(namespace string, iface *NetworkInterface) error
 		return err
 	}
 	return nil
+}
+
+// ConfigureInterface sets up the interfaces with the correct name, network namesapce and ip address
+func (me *Cni) ConfigureInterface(cmd util.CniCommand, result *currentcni.Result) error {
+
+	mac := result.Interfaces[0].Mac
+	ifaceName, err := GetIfaceNameByMac(mac)
+	if err != nil {
+		return err
+	}
+
+	iface := &NetworkInterface{
+		Name:     ifaceName,
+		DestName: result.Interfaces[0].Name,
+		Address:  &result.IPs[0].Address,
+	}
+
+	err = me.nw.Configure(cmd.Netns, iface)
+	if err != nil {
+		return fmt.Errorf("failed to configure interface %w", err)
+	}
+	return err
+}
+
+// GetIfaceNameByMac returns the name of an interface matching the given MAC address
+func GetIfaceNameByMac(mac string) (string, error) {
+	ifaces, err := net.Interfaces()
+	if err != nil {
+		return "", err
+	}
+
+	for _, iface := range ifaces {
+		if iface.HardwareAddr.String() == mac {
+			return iface.Name, nil
+		}
+	}
+
+	return "", fmt.Errorf("failed to find interface for %s", mac)
 }
